\name{SniperFunctions}
\Rdversion{1.0.0}
\alias{SniperFunctions}
\docType{package}
\title{SniperFunctions}
\description{SniperFunctions}
\details{
\tabular{ll}{
Package: \tab SniperFunctions\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2011-01-05\cr
License: \tab	What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
KS
Maintainer: Who to complain to <ks@me.com>
}
\references{}
\seealso{
\code{\link{trend_monitor}}
}

